refactor(app): separate auth logic from system logic in different pages component
