+++ src/enums/BudgetTableAction.enum.ts
@@ -2,4 +2,5 @@ export enum BudgetTableActionEnum {
   EDIT = "edit",
   DELETE = "delete",
   CREATE = "create",
+  NONE = "none",
 }
